Ajustando tipo id do alerta

+++ src/main/java/com/panicButton/panicButton/proxy/IProxy.java
@@ -5,10 +5,10 @@ import com.panicButton.panicButton.domain.Usuario;
 
 public interface IProxy {
     Usuario createUsuario(Usuario usuario);
-    Alerta getAlerta(int id);
-    void removeAlerta(int id);
+    Alerta getAlerta(Long id);
+    void removeAlerta(Long id);
     Alerta createAlerta(Alerta alerta);
-    Alerta updateAlerta(int id, Alerta alerta);
+    Alerta updateAlerta(Long id, Alerta alerta);
     Usuario getUsuario(String matricula);
     Usuario updateUsuario(String matricula, Usuario usuario);
 }
